Move `WeakRefVirtualProxy` to it's own file
