+++ EssentialFeed/EssentialFeediOS/Feed UI/Composers/FeedUIComposer.swift
@@ -41,20 +41,6 @@ private extension FeedViewController {
     }
 }
 
-private final class WeakRefVirtualProxy<T: AnyObject> {
-    private weak var object: T?
-    
-    init(_ object: T) {
-        self.object = object
-    }
-}
-
-extension WeakRefVirtualProxy: FeedLoadingView where T:FeedLoadingView {
-    func display(_ viewModel: FeedLoadingViewModel) {
-        object?.display(viewModel)
-    }
-}
-
 private final class FeedViewAdapter: FeedView {
     private weak var feedViewController: FeedViewController?
     private var imageLoader: FeedImageDataLoader
@@ -75,12 +61,6 @@ private final class FeedViewAdapter: FeedView {
             
             return view
         }
-    }
-}
-
-extension WeakRefVirtualProxy: FeedImageView where T:FeedImageView, T.Image == UIImage {
-    func display(_ model: FeedImageViewModel<UIImage>) {
-        object?.display(model)
     }
 }
